# Conflicts:
#	src/main/java/starlight/backend/proof/ProofController.java
#	src/main/java/starlight/backend/proof/service/ProofServiceInterface.java

+++ src/main/java/starlight/backend/talent/service/impl/TalentServiceImpl.java
@@ -2,13 +2,17 @@ package starlight.backend.talent.service.impl;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
-import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 import starlight.backend.exception.PageNotFoundException;
 import starlight.backend.exception.TalentNotFoundException;
+import starlight.backend.security.service.SecurityServiceInterface;
 import starlight.backend.talent.MapperTalent;
 import starlight.backend.talent.model.request.TalentUpdateRequest;
 import starlight.backend.talent.model.response.TalentFullInfo;
@@ -24,16 +28,16 @@ import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Service
+@Transactional
 public class TalentServiceImpl implements TalentServiceInterface {
-    MapperTalent mapper;
-    UserRepository repository;
-    PositionRepository positionRepository;
-
+    private MapperTalent mapper;
+    private UserRepository repository;
+    private PositionRepository positionRepository;
+    private SecurityServiceInterface securityService;
     @PersistenceContext
     private EntityManager em;
 
     @Override
-    @Transactional
     public TalentPagePagination talentPagination(int page, int size) {
         var pageRequest = repository.findAll(
                 PageRequest.of(page, size, Sort.by("userId").descending())
@@ -44,7 +48,6 @@ public class TalentServiceImpl implements TalentServiceInterface {
     }
 
     @Override
-    @Transactional
     public Optional<TalentFullInfo> talentFullInfo(long id) {
         return Optional.of(repository.findById(id)
                 .map(mapper::toTalentFullInfo)
@@ -52,8 +55,10 @@ public class TalentServiceImpl implements TalentServiceInterface {
     }
 
     @Override
-    @Transactional
-    public TalentFullInfo updateTalentProfile(long id, TalentUpdateRequest talentUpdateRequest) {
+    public TalentFullInfo updateTalentProfile(long id, TalentUpdateRequest talentUpdateRequest,Authentication auth) {
+        if (securityService.checkingLoggedAndToken(id, auth)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
         return repository.findById(id).map(talent -> {
             talent.setFullName(talentUpdateRequest.fullName());
             talent.setBirthday(talentUpdateRequest.birthday());
@@ -72,8 +77,10 @@ public class TalentServiceImpl implements TalentServiceInterface {
     }
 
     @Override
-    @Transactional
-    public void deleteTalentProfile(long talentId) {
+    public void deleteTalentProfile(long talentId, Authentication auth) {
+        if (securityService.checkingLoggedAndToken(talentId, auth)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
         UserEntity user = em.find(UserEntity.class, talentId);
         user.setPositions(null);
         em.remove(user);
